Update reddit_email_bot.py
I changed the search criteria from being a list of subreddits and a list of words/phrases to a dictionary. This should allow the user set different keywords/phrases per subreddit. 
Also some minor bug fixes in the e-mail portion of the script to handle for the maximum character length and a bug that was coming up with certain characters.

Ran this for a while this morning, works well except occasionally the encode doesn't know a character and yields "???", and I also got 2 or 3 blank messages (out of probably 200), not sure what that is about.